requirements.txt
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
PyAutoGUI==0.9.54
pyperclip==1.9.0
pytesseract==0.3.13
vertexai==1.71.1

chess~=1.11.2
ImageHash~=4.3.2